Bug #767: Problems making Squid 3.0 PRE3 64 bit
Fixed the following on HP Tru64 and probably on some other 64 bit platforms:

- assert() must test logical expressions, not pointers
- STATUS define conflict in parse.c (snmplib)
- Warnings SMB, fakeauth and MSNT helpers
- Various debug() ==> debugs() changes
- Alpha asm access to CPU profiling info
- timezone should be long, not time_t
- Various clean of ESI code
- assert() definition problem in diskd.cc

+++ src/ESICustomParser.h
@@ -1,5 +1,5 @@
 /*
- * $Id: ESICustomParser.h,v 1.5 2005/03/28 21:44:12 hno Exp $
+ * $Id: ESICustomParser.h,v 1.6 2005/07/03 15:25:08 serassio Exp $
  *
  *
  * SQUID Web Proxy Cache          http://www.squid-cache.org/
@@ -45,7 +45,7 @@ public:
     ~ESICustomParser();
     /* true on success */
     bool parse(char const *dataToParse, size_t const lengthOfData, bool const endOfStream);
-    size_t lineNumber() const;
+    long int lineNumber() const;
     char const * errorString() const;
 
 private:
